Added JSON Utilities
Added utility class `JsonUtil` to JSON support serialization and manipulation.
Updated class `SvcTable` to be iterable in range based loops.
Added field `proto` to class `Record`.
Added member `proto` to enumeration type `SvcField`.
Other small changes including refactoring and reformatting.

+++ SvcScan/src/includes/io/filesys/path_info.h
@@ -13,7 +13,7 @@ namespace scan
     /**
     * @brief  File path information enumeration type.
     */
-    enum class PathInfo : unsigned short
+    enum class PathInfo : unsigned int
     {
         unknown,          // Unknown file path information
         exists,           // Existing file path
